# Fix incremental compilation, again