--- a/test/IncrementalCompilation/src/IncrementalCompilation.jl
+++ b/test/IncrementalCompilation/src/IncrementalCompilation.jl
@@ -1,0 +1,10 @@
+__precompile__()
+module IncrementalPrecompilation
+
+using NamedTuples
+
+function demo()
+    @NT(cols=5)
+end
+
+end