The DataStreams package was running into issues with incremental pre-compilation with the NamedTuples package:

```julia
julia> Base.compilecache("DataStreams")
INFO: Recompiling stale cache file /Users/omus/.julia/lib/v0.6/DataStreams.ji for module DataStreams.
WARNING: eval from module NamedTuples to Data:
Expr(:type, false, Expr(:<:, Expr(:curly, :_NT_col, :T1)::Any, NamedTuples.NamedTuple)::Any, Expr(:block, Expr(:::, :col, :T1)::Any, Expr(:tuple)::Any)::Any)::Any
  ** incremental compilation may be broken for this module **

WARNING: eval from module NamedTuples to Data:
Expr(:type, false, Expr(:<:, Expr(:curly, :_NT_name_compute_computeargs, :T1, :T2, :T3)::Any, NamedTuples.NamedTuple)::Any, Expr(:block, Expr(:::, :name, :T1)::Any, Expr(:::, :compute, :T2)::Any, Expr(:::, :computeargs, :T3)::Any, Expr(:tuple)::Any)::Any)::Any
  ** incremental compilation may be broken for this module **
```

These warnings occurred because the current implementation of `@NT` performs an `eval` during macro expansion. In this PR the `@NT` macro has been revised to only perform the `eval` at runtime which makes incremental compilation work. To do this I needed to break apart the `make_tuple` function into two parts: initial expression parsing and type creation later. In order to make this change completely non-breaking I have left `make_tuple` as is and placed the initial expression parsing components directly in the `@NT` macro and have put the runtime type creation part of the `NT` function.

These changes should properly address the change that @quinnj wanted as part of https://github.com/JuliaData/NamedTuples.jl/pull/46 without generating duplicate types and should allow the DataStreams package to no longer require a custom fork of NamedTuples.jl