@omus 2018-05-07
Note: Originally this was introduced and merged as https://github.com/JuliaData/NamedTuples.jl/pull/61. Unfortunately the original implementation introduced a performance hit when it was merged so it was reverted. This PR has improved performance but it has been improved from the original PR.

Before this PR:
```julia
julia> @benchmark @NT(a=2, b=4)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     6.522 ns (0.00% GC)
  median time:      7.334 ns (0.00% GC)
  mean time:        9.052 ns (0.00% GC)
  maximum time:     2.305 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999
```

With this PR:
```julia
julia> @benchmark @NT(a=2, b=4)
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  1
  --------------
  minimum time:     223.069 ns (0.00% GC)
  median time:      255.001 ns (0.00% GC)
  mean time:        329.682 ns (0.51% GC)
  maximum time:     17.416 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     478
```
For the initial execution of `@NT(a=2, b=4)` and `eval` is called at runtime to generate the type. For subsequent calls only a `isdefined` and `getfield` call occur.