@omus 2018-05-07
I'm concerned that the benchmarks shown above may not be accurate as separating the macro and runtime as follows makes the timing be almost identical.

Before this PR:
```julia
julia> T = @NT(a, b)
NamedTuples._NT_a_b

julia> @benchmark $T(2, 4)
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  1
  --------------
  minimum time:     207.980 ns (0.00% GC)
  median time:      252.435 ns (0.00% GC)
  mean time:        350.668 ns (0.76% GC)
  maximum time:     42.068 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     550
```

With this PR:
```julia
julia> T = @NT(a, b)
NamedTuples._NT_a_b

julia> @benchmark $T(2, 4)
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  1
  --------------
  minimum time:     202.557 ns (0.00% GC)
  median time:      239.580 ns (0.00% GC)
  mean time:        306.513 ns (0.62% GC)
  maximum time:     12.167 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     530
```