@davidanthoff 2018-05-07
I tried this:
````julia
function foo(x)
    a = @NT(a=x,b=5)
    return a.a + a.b
end
````
Then I get the following timings.
Pre-this PR:
````
julia> @benchmark foo(3)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.099 ns (0.00% GC)
  median time:      2.300 ns (0.00% GC)
  mean time:        2.336 ns (0.00% GC)
  maximum time:     24.000 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
````
With this PR:
````
julia> @benchmark foo(3)
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  1
  --------------
  minimum time:     279.078 ns (0.00% GC)
  median time:      304.252 ns (0.00% GC)
  mean time:        324.123 ns (0.54% GC)
  maximum time:     5.176 μs (90.64% GC)
  --------------
  samples:          10000
  evals/sample:     282
````
I've also just tried this PR with a run of the mill [Query.jl](https://github.com/davidanthoff/Query.jl) query, and it is about a factor 40 slower for the one example I tried.

So I do think that this PR has major performance regressions. I think broadly speaking we really can't put anything extra into the construction phase: in almost all the use cases that I'm aware off (Queryverse, DataStreams [although I'm least familiar with that, so take that with a grain of salt], JuliaDB), named tuple construction sits in the inner most loop, and adding even a small overhead just kills performance.