@quinnj 2018-06-05
@omus, what's the status here? I'm sympathetic to switching DataStreams back over to using NamedTuples, but the circular dependencies of PRs between here & DataStreams are pretty difficult to follow; I'm not sure which PRs are current or where I should start reviewing.