@davidanthoff 2018-06-05
I think this PR here should be closed and not merged. I think the strategy here has too much of a performance impact.

I think to fix the precompile issue in DataStreamsjl, something like https://github.com/JuliaData/DataStreams.jl/pull/74 is more promising because it doesn't impact all the other packages that rely on NamedTuples.jl.