@omus 2018-06-05
I agree the performance impact of this is problematic but this is probably the best general approach for fixing incremental pre-compilation for other packages. My opinion is we should hold off on merging this until @davidanthoff has ported most/all of his packages to Julia 0.7 and then we could merge this for those packages wanting to maintain Julia 0.6 and 0.7 compatibility.